requirements.txt
pandas
xmltodict
google-cloud-bigquery
google-cloud-storage
gvm-tools
schedule
requests
